PrimaryKeyRelatedField in serializers.

+++ backend/profiles/models.py
@@ -26,7 +26,7 @@ class InstructorProfile(BaseModel):
         on_delete=models.CASCADE
     )
     date_of_birth = models.DateField()
-    experience = models.IntegerField()
+    experience = models.DecimalField(max_digits=3, decimal_places=1)
 
     def __str__(self):
         return self.user.username
